Adaptation du projet aux clouds, la classe EndpointAsycTask a été modifié pour n'agir que sur les User

+++ servlets-backend/src/main/java/servlets/backend/Choice.java
@@ -15,7 +15,7 @@ public class Choice {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String description;
     private int idExercice;
     private String choice1;
@@ -32,7 +32,7 @@ public class Choice {
         this.choice3 = choice3;
     }
 
-    public Choice(int id, String description, String choice1, String choice2, String choice3)
+    public Choice(long id, String description, String choice1, String choice2, String choice3)
     {
         this.description = description;
         this.id = id;
@@ -42,11 +42,11 @@ public class Choice {
     }
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
